add jpeg to filters and refactor upload (WIP)
