+++ lumina/components/FollowerFeed.tsx
@@ -32,7 +32,7 @@ const FollowerFeed: React.FC<FollowerFeedProps> = ({ pubkey }) => {
 
   // const filteredEvents = events.filter((event) => event.content.includes(".jpg"));
   // filter events with regex that checks for png, jpg, or gif
-  let filteredEvents = events.filter((event) => event.content.match(/https?:\/\/.*\.(?:png|jpg|gif|mp4|webm)/g)?.[0]);
+  let filteredEvents = events.filter((event) => event.content.match(/https?:\/\/.*\.(?:png|jpg|gif|mp4|webm|jpeg)/g)?.[0]);
 
   // now filter all events with a tag[0] == t and tag[1] == nsfw
   filteredEvents = filteredEvents.filter((event) => event.tags.map((tag) => tag[0] == "t" && tag[1] == "nsfw"));
