+++ lumina/components/GlobalFeed.tsx
@@ -16,7 +16,7 @@ const GlobalFeed: React.FC = () => {
 
   // const filteredEvents = events.filter((event) => event.content.includes(".jpg"));
   // filter events with regex that checks for png, jpg, or gif
-  let filteredEvents = events.filter((event) => event.content.match(/https?:\/\/.*\.(?:png|jpg|gif)/g)?.[0]);
+  let filteredEvents = events.filter((event) => event.content.match(/https?:\/\/.*\.(?:png|jpg|gif|jpeg)/g)?.[0]);
 
   // now filter all events with a tag[0] == t and tag[1] == nsfw
   // filteredEvents = filteredEvents.filter((event) => event.tags.map((tag) => tag[0] == "t" && tag[1] == "nsfw"));
