+++ lumina/components/TrendingImage.tsx
@@ -41,8 +41,8 @@ const TrendingImage: React.FC<TrendingImageProps> = ({ eventId, pubkey }) => {
   const createdAt = events && events.length > 0 ? new Date(events[0].created_at * 1000) : new Date();
   const title = userData?.username || userData?.display_name || userData?.name || userData?.npub || npubShortened;
   text = text.replaceAll('\n', ' ');
-  const imageSrc = text.match(/https?:\/\/[^ ]*\.(png|jpg|gif)/g);
-  const textWithoutImage = text.replace(/https?:\/\/.*\.(?:png|jpg|gif)/g, '');
+  const imageSrc = text.match(/https?:\/\/[^ ]*\.(png|jpg|gif|jpeg)/g);
+  const textWithoutImage = text.replace(/https?:\/\/.*\.(?:png|jpg|gif|jpeg)/g, '');
   const hrefProfile = `/profile/${nip19.npubEncode(pubkey)}`;
   const profileImageSrc = userData?.picture || "https://robohash.org/" + pubkey;
